QRPRINTER/requirements.txt
Flask>=3.0.0
qrcode[pil]>=7.4.2
Pillow>=10.2.0
watchdog>=3.0.0
requests>=2.31.0

